renamed cltk.phonology.old_norse -> cltk.phonology.non, improved docs and first draft for Old Norse phonology processes
