+++ src/cltk/phonology/non/syllabifier.py
@@ -108,3 +108,17 @@ ipa_hierarchy = [
     ["f", "v", "θ", "ð", "s", "h"],
     ["b", "d", "g", "k", "p", "t"],
 ]
+
+
+class OldNorseSyllabifier:
+    def __init__(self):
+        pass
+
+    def syllabify(self, word):
+        pass
+
+    def __repr__(self):
+        return f"<OldNorseScanner>"
+
+    def __call__(self, word):
+        return self.syllabify(word)
