+++ src/cltk/phonology/non/transcription.py
@@ -6,7 +6,7 @@ Sources:
 
 from typing import Union
 
-from cltk.phonology.old_norse.syllabifier import BACK_TO_FRONT_VOWELS
+from cltk.phonology.non.syllabifier import BACK_TO_FRONT_VOWELS
 from cltk.phonology.utils import (
     AbstractConsonant,
     AbstractPosition,
