+++ src/cltk/phonology/syllabify.py
@@ -7,8 +7,8 @@ from cltk.core.exceptions import CLTKException
 from cltk.phonology.middle_english.syllabifier import Syllabifier as ME_Syllabifier
 from cltk.phonology.middle_high_german.syllabifier import Syllabifier as MHG_Syllabifier
 from cltk.phonology.old_english.syllabifier import Syllabifier as OE_Syllabifier
-from cltk.phonology.old_norse.syllabifier import hierarchy as old_norse_hierarchy
-from cltk.phonology.old_norse.syllabifier import (
+from cltk.phonology.non.syllabifier import hierarchy as old_norse_hierarchy
+from cltk.phonology.non.syllabifier import (
     ipa_hierarchy as ipa_old_norse_hierarchy,
 )
